moved container_adapter_client.ts to root of test
